pyproject.toml
[project]
name = "cinema"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.2",
    "python-decouple>=3.8",
    "dj-database-url>=2.3",
    "pillow>=11.0",
    "djangorestframework-simplejwt>=5.5.1",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "django-extensions>=3.2",
    "django-debug-toolbar>=4.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
